refactor(tapir-processing): :art: Moved process_video logic into VideoProcessor
