--- src/backend/point_cloud/point_cloud_type.py
@@ -1,5 +1,0 @@
-from enum import Enum
-
-
-class PointCloudType(Enum):
-    TAPIR = 1
